# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NQL DZ Transport Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /driver/dashboard:
    get:
      operationId: getDriverDashboard
      tags: [driver]
      summary: Get driver dashboard data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Driver dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDashboard"
  /driver/scan:
    post:
      operationId: driverScanCard
      tags: [driver]
      summary: Scan customer QR code and deduct fare
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Scan successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "400":
          description: Insufficient balance or invalid card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /driver/trips:
    get:
      operationId: getDriverTrips
      tags: [driver]
      summary: Get driver trips list
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Trips list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverTripsResponse"
  /driver/report/daily:
    get:
      operationId: getDriverDailyReport
      tags: [driver]
      summary: Get driver daily report data
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverDailyReport"
  /driver/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [driver]
      summary: Request withdrawal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequestCreate"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequest"
        "400":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /driver/withdrawals:
    get:
      operationId: getDriverWithdrawals
      tags: [driver]
      summary: Get driver withdrawal history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Withdrawal history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequestsListResponse"

  # CUSTOMER
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        cardId:
          type: integer
        driverId:
          type: integer
          nullable: true
        distributorId:
          type: integer
          nullable: true
        amount:
          type: number
        platformFee:
          type: number
        driverEarning:
          type: number
        type:
          type: string
          enum: [ride, topup]
        cardType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - cardId
        - amount
        - platformFee
        - driverEarning
        - type
        - createdAt
...
    ScanRequest:
      type: object
      properties:
        cardNumber:
          type: string
      required:
        - cardNumber
...
    ScanResult:
      type: object
      properties:
        success:
          type: boolean
        cardType:
          type: string
        fareAmount:
          type: number
        platformFee:
          type: number
        driverEarning:
          type: number
        cardBalance:
          type: number
        message:
          type: string
      required:
        - success
        - cardType
        - fareAmount
        - platformFee
        - driverEarning
        - cardBalance
        - message
    DriverDashboard:
      type: object
      properties:
        balance:
          type: number
        todayPassengers:
          type: integer
        todayEarnings:
          type: number
        platformFeeToday:
          type: number
        totalTrips:
          type: integer
        recentTrips:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - balance
        - todayPassengers
        - todayEarnings
        - platformFeeToday
        - totalTrips
        - recentTrips
    DriverTripsResponse:
      type: object
      properties:
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        totalEarnings:
          type: number
      required:
        - trips
        - total
        - totalEarnings
    DriverDailyReport:
      type: object
      properties:
        date:
          type: string
        totalTrips:
          type: integer
        totalEarnings:
          type: number
        platformFee:
          type: number
        passengers:
          type: integer
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - date
        - totalTrips
        - totalEarnings
        - platformFee
        - passengers
        - trips
    WithdrawalRequest:
      type: object
      properties:
        id:
          type: integer
        driverId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, ccp]
        phone:
          type: string
          nullable: true
        ccpAccount:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - driverId
        - amount
        - method
        - status
        - createdAt
    WithdrawalRequestsListResponse:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/WithdrawalRequest"
        total:
          type: integer
      required:
        - requests
        - total
    WithdrawalRequestCreate:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, ccp]
        phone:
          type: string
          nullable: true
        ccpAccount:
          type: string
          nullable: true
        holderName:
          type: string
        holderLastName:
          type: string
      required:
        - amount
        - method
        - holderName
        - holderLastName
...