# lib/api-zod/src/generated/types/driverDailyReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NQL DZ Transport Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface DriverDailyReport {
  date: string;
  totalTrips: number;
  totalEarnings: number;
  platformFee: number;
  passengers: number;
  trips: Transaction[];
}