# lib/api-zod/src/generated/types/driverDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NQL DZ Transport Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface DriverDashboard {
  balance: number;
  todayPassengers: number;
  todayEarnings: number;
  platformFeeToday: number;
  totalTrips: number;
  recentTrips: Transaction[];
}